Northgayd bot/requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
discord.py==1.5.1
dnspython==2.0.0
idna==2.10
multidict==4.7.6
pip==20.2.3
pymongo==3.11.0
setuptools==49.2.1
yarl==1.5.1